backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pdfminer.six>=20221105
pytesseract>=0.3.10
Pillow>=10.3.0
ollama>=0.2.1
pydantic>=2.7.0
